requirements.txt
streamlit==1.23.1
torch>=2.0.0,<3.0.0
transformers>=4.30.2,<5.0.0
sentence-transformers>=2.2.2,<3.0.0
pdfplumber==0.11.7
faiss-cpu>=1.12.0,<2.0.0
bitsandbytes==0.39.0
accelerate>=0.22.0,<1.0.0
numpy>=1.24.0,<1.27.0
